Fix jest test suite by setting up babel-jest
Oops. Turns out we actually need to use babel in this project. We don't
need to use it to build the Chrome extension, BUT we DO need to use it
to run the tests correctly in the node environment. This restores
jest and configures it to only run for jest with the node preset.